+++ golang/lib/client/do_nothing_client/do_nothing_client.go
@@ -1,0 +1,16 @@
+package do_nothing_client
+
+import "github.com/sirupsen/logrus"
+
+type DoNothingClient struct {
+
+}
+
+func NewDoNothingClient() *DoNothingClient {
+	return &DoNothingClient{}
+}
+
+func (client DoNothingClient) TrackUserAcceptSendingMetrics(userAcceptSendingMetrics bool) error {
+	logrus.Debugf("Do nothing client TrackUserAcceptSendingMetrics called")
+	return nil
+}