+++ golang/lib/client/metrics_client_creator.go
@@ -1,17 +1,21 @@
 package client
 
 import (
+	"github.com/kurtosis-tech/metrics-library/golang/lib/client/do_nothing_client"
 	"github.com/kurtosis-tech/metrics-library/golang/lib/client/snow_plow_client"
 	"github.com/kurtosis-tech/metrics-library/golang/lib/source"
 	"github.com/kurtosis-tech/stacktrace"
 )
 
-const (
-	defaultMetricsClientProvider = SnowPlow
-)
+func CreateDefaultMetricsClient(source source.Source, userId string, usserAcceptSendingMetrics bool) (MetricsClient, error) {
 
-func CreateDefaultMetricsClient(source source.Source, userId string) (MetricsClient, error) {
-	return CreateMetricsClient(source, userId, defaultMetricsClientProvider)
+	metricsProvider := DoNoting
+
+	if usserAcceptSendingMetrics{
+		metricsProvider = SnowPlow
+	}
+
+	return CreateMetricsClient(source, userId, metricsProvider)
 }
 
 func CreateMetricsClient(source source.Source, userId string, metricsProvider MetricsClientProvider) (MetricsClient, error) {
@@ -22,6 +26,9 @@ func CreateMetricsClient(source source.Source, userId string, metricsProvider Me
 		if err != nil {
 			return nil, stacktrace.Propagate(err, "An error occurred creating SnowPlow metrics client")
 		}
+		return metricsClient, nil
+	case DoNoting:
+		metricsClient := do_nothing_client.NewDoNothingClient()
 		return metricsClient, nil
     default:
 		return nil, stacktrace.NewError("Unrecognized metrics provider '%v'", metricsProvider)